Ultimo 6-01-2026

+++ quality_control/models/stock_picking.py
@@ -2,6 +2,7 @@
 
 from odoo import fields, models, _
 from odoo.exceptions import UserError
+from odoo.tools.misc import OrderedSet
 from odoo.tools.float_utils import float_is_zero
 
 
@@ -18,7 +19,10 @@ class StockPicking(models.Model):
         for picking in self:
             todo = False
             fail = False
-            for check in picking.check_ids:
+            # Only prefetch needed QC fields to avoid to bloat the cache by fetching other QC data.
+            checks = picking.check_ids
+            checks.fetch(['quality_state'])
+            for check in checks:
                 if check.quality_state == 'none':
                     todo = True
                 elif check.quality_state == 'fail':
@@ -40,20 +44,29 @@ class StockPicking(models.Model):
         for picking in self:
             picking.quality_alert_count = len(picking.quality_alert_ids)
 
+    def _checks_to_do(self):
+        check_ids_to_do = OrderedSet()
+        for picking in self:
+            has_picked = True
+            if all(not move.picked for move in picking.move_ids):
+                checkable_lines = picking.move_line_ids
+                has_picked = False
+            else:
+                checkable_lines = picking.move_line_ids.filtered(
+                    lambda ml: ml._is_checkable(check_picked=has_picked)
+                )
+            checkable_products = checkable_lines.product_id
+            checks_to_do = self.check_ids.filtered(
+                lambda qc: qc._is_to_do(checkable_products, check_picked=has_picked)
+            )
+            check_ids_to_do.update(checks_to_do.ids)
+        return self.env['quality.check'].browse(check_ids_to_do)
+
     def check_quality(self):
-        if all(not move.picked for move in self.move_ids):
-            checkable_lines = self.move_line_ids
-        else:
-            checkable_lines = self.move_line_ids.filtered(
-            lambda ml: (
-                (not self.env.context.get('picking_validation') or ml.move_id.picked) and
-                not float_is_zero(ml.quantity, precision_rounding=ml.product_uom_id.rounding)
-            ))
-        checkable_products = checkable_lines.product_id
-        checks = self.check_ids.filtered(lambda check: check.quality_state == 'none' and (check.product_id in checkable_products or check.measure_on == 'operation'))
+        checks = self._checks_to_do()
         if checks:
             return checks.action_open_quality_check_wizard()
-        return False
+        return True
 
     def _create_backorder(self, backorder_moves=None):
         res = super(StockPicking, self)._create_backorder(backorder_moves=backorder_moves)
@@ -62,7 +75,8 @@ class StockPicking(models.Model):
         for backorder in res:
             # Do not link the QC of move lines with quantity of 0 in backorder.
             backorder.move_line_ids.filtered(lambda ml: not float_is_zero(ml.quantity, precision_rounding=ml.product_uom_id.rounding)).check_ids.picking_id = backorder
-            backorder.backorder_id.check_ids.filtered(lambda qc: qc.quality_state == 'none').sudo().unlink()
+            if backorder.backorder_id.state in ('done', 'cancel'):
+                backorder.backorder_id.check_ids.filtered(lambda qc: qc.quality_state == 'none').sudo().unlink()
             backorder.move_ids._create_quality_checks()
         return res
 
@@ -74,16 +88,13 @@ class StockPicking(models.Model):
     def _pre_action_done_hook(self):
         res = super()._pre_action_done_hook()
         if res is True:
-            pickings_to_check_quality = self._check_for_quality_checks()
-            if pickings_to_check_quality:
-                return pickings_to_check_quality.with_context(picking_validation=True, pickings_to_check_quality=pickings_to_check_quality.ids).check_quality()
+            return self.with_context(picking_validation=True).check_quality()
         return res
 
     def _check_for_quality_checks(self):
         quality_pickings = self.env['stock.picking']
         for picking in self:
-            product_to_check = picking.mapped('move_line_ids').filtered(lambda ml: ml.picked).mapped('product_id')
-            if picking.mapped('check_ids').filtered(lambda qc: qc.quality_state == 'none' and (qc.product_id in product_to_check or qc.measure_on == 'operation')):
+            if picking._checks_to_do():
                 quality_pickings |= picking
         return quality_pickings
